ENG-7284 - Refactor Datastore pt 1. Update dependency injections to use same instance and saveEvent method.
