+++ NeuroID/src/main/java/com/neuroid/tracker/callbacks/NIDContextMenuCallbacks.kt
@@ -8,10 +8,11 @@ import com.neuroid.tracker.events.PASTE
 import com.neuroid.tracker.events.COPY
 import com.neuroid.tracker.events.CUT
 import com.neuroid.tracker.models.NIDEventModel
-import com.neuroid.tracker.storage.getDataStoreInstance
+import com.neuroid.tracker.storage.NIDDataStoreManager
 import com.neuroid.tracker.utils.NIDLog
 
 abstract class NIDContextMenuCallBacks(
+    val dataStore: NIDDataStoreManager,
     actionCallBack: ActionMode.Callback?
 ) : ActionMode.Callback {
     val wrapper = actionCallBack
@@ -29,12 +30,37 @@ abstract class NIDContextMenuCallBacks(
     override fun onDestroyActionMode(actionMode: ActionMode?) {
         wrapper?.onDestroyActionMode(actionMode)
     }
+
+    internal fun saveEvent(option: Int, item: String) {
+        val type = when (option) {
+            android.R.id.paste -> PASTE
+            android.R.id.copy -> COPY
+            android.R.id.cut -> CUT
+            else -> ""
+        }
+
+        if (type.isNotEmpty()) {
+            this.dataStore.saveEvent(
+                NIDEventModel(
+                    type = CONTEXT_MENU,
+                    attrs = listOf(
+                        mapOf(
+                            "option" to "$option",
+                            "item" to item,
+                            "type" to type
+                        )
+                    )
+                )
+            )
+        }
+    }
 }
 
 // This is the callback for the context menu that appears when text is already in field
 class NIDTextContextMenuCallbacks(
+    dataStore: NIDDataStoreManager,
     actionCallBack: ActionMode.Callback?
-) : NIDContextMenuCallBacks(actionCallBack) {
+) : NIDContextMenuCallBacks(dataStore, actionCallBack) {
 
     override fun onActionItemClicked(action: ActionMode?, item: MenuItem?): Boolean {
         wrapper?.onActionItemClicked(action, item)
@@ -50,8 +76,9 @@ class NIDTextContextMenuCallbacks(
 
 // This is the callback for the context menu that appears when the text field is empty (only available in later API versions)
 class NIDLongPressContextMenuCallbacks(
+    dataStore: NIDDataStoreManager,
     actionCallBack: ActionMode.Callback?
-) : NIDContextMenuCallBacks(actionCallBack) {
+) : NIDContextMenuCallBacks(dataStore, actionCallBack) {
     override fun onActionItemClicked(action: ActionMode?, item: MenuItem?): Boolean {
         wrapper?.onActionItemClicked(action, item)
 
@@ -64,33 +91,3 @@ class NIDLongPressContextMenuCallbacks(
     }
 }
 
-private fun saveEvent(option: Int, item: String) {
-    val gyroData = NIDSensorHelper.getGyroscopeInfo()
-    val accelData = NIDSensorHelper.getAccelerometerInfo()
-
-    val type = when (option) {
-        android.R.id.paste -> PASTE
-        android.R.id.copy -> COPY
-        android.R.id.cut -> CUT
-        else -> ""
-    }
-
-    if (type.isNotEmpty()) {
-        getDataStoreInstance()
-            .saveEvent(
-                NIDEventModel(
-                    type = CONTEXT_MENU,
-                    ts = System.currentTimeMillis(),
-                    gyro = gyroData,
-                    accel = accelData,
-                    attrs = listOf(
-                        mapOf(
-                            "option" to "$option",
-                            "item" to item,
-                            "type" to type
-                        )
-                    )
-                )
-            )
-    }
-}
