+++ NeuroID/src/main/java/com/neuroid/tracker/utils/JsonUtils.kt
@@ -5,19 +5,16 @@ import com.neuroid.tracker.extensions.getSHA256withSalt
 class JsonUtils {
     companion object {
         fun getAttrJson(text: String?): List<Map<String, String>> {
-
-
             val value =
                 mapOf(
                     "n" to "v",
                     "v" to "S~C~~${text?.length ?: 0}"
                 )
 
-            val hashValue = text?.getSHA256withSalt()?.take(8)
             val hash =
                 mapOf(
                     "n" to "hash",
-                    "v" to if (hashValue == null) {""} else {hashValue}
+                    "v" to (text?.getSHA256withSalt()?.take(8) ?: "")
                 )
 
             return listOf(
