+++ NeuroID/src/main/java/com/neuroid/tracker/utils/NIDTextWatcher.kt
@@ -3,14 +3,16 @@ package com.neuroid.tracker.utils
 import android.text.Editable
 import android.text.TextWatcher
 import com.neuroid.tracker.NeuroID
-import com.neuroid.tracker.callbacks.NIDSensorHelper
 import com.neuroid.tracker.events.INPUT
 import com.neuroid.tracker.events.PASTE
 import com.neuroid.tracker.extensions.getSHA256withSalt
 import com.neuroid.tracker.models.NIDEventModel
+import com.neuroid.tracker.storage.NIDDataStoreManager
 import com.neuroid.tracker.utils.JsonUtils.Companion.getAttrJson
 
 class NIDTextWatcher(
+    val dataStore: NIDDataStoreManager,
+    val logger: NIDLogWrapper,
     private val idName: String,
     val className: String? = "",
     val startingHashValue: String? = ""
@@ -25,7 +27,6 @@ class NIDTextWatcher(
     }
 
     override fun onTextChanged(sequence: CharSequence?, start: Int, before: Int, count: Int) {
-
         // Check if the change is due to a paste operation
         val clipboard = NeuroID.getInstance()?.getClipboardManagerInstance()
         val clipData = clipboard?.primaryClip
@@ -35,7 +36,7 @@ class NIDTextWatcher(
                 pastedText = clipData.getItemAt(0).text.toString()
             } catch (e: Exception) {
                 e.message?.let {
-                    NIDLog.e("Activity", it)
+                    logger.e("Activity", it)
                 }
             }
             val pasteCount = pastedText.length
@@ -46,16 +47,11 @@ class NIDTextWatcher(
                 // Checks if paste operation is duplicated ENG-6236
                 if (currentPastedHashValue != lastPastedHashValue) {
                     lastPastedHashValue = sequence?.toString()?.hashCode().toString()
-                    val ts = System.currentTimeMillis()
-                    val gyroData = NIDSensorHelper.getGyroscopeInfo()
-                    val accelData = NIDSensorHelper.getAccelerometerInfo()
 
                     if (pastedText.isNotEmpty()) {
-                        NeuroID.getInstance()?.dataStore
-                            ?.saveEvent(
+                        dataStore.saveEvent(
                                 NIDEventModel(
                                     type = PASTE,
-                                    ts = ts,
                                     tg = hashMapOf(
                                         "attr" to getAttrJson(sequence.toString()),
                                         "et" to "text"
@@ -63,8 +59,6 @@ class NIDTextWatcher(
                                     tgs = idName,
                                     v = "S~C~~${sequence?.length}",
                                     hv = sequence?.toString()?.getSHA256withSalt()?.take(8),
-                                    gyro = gyroData,
-                                    accel = accelData,
                                     attrs = listOf(
                                         mapOf(
                                             "clipboardText" to "S~C~~${pastedText.length}"
@@ -80,20 +74,15 @@ class NIDTextWatcher(
     }
 
     override fun afterTextChanged(sequence: Editable?) {
-        val ts = System.currentTimeMillis()
-        val gyroData = NIDSensorHelper.getGyroscopeInfo()
-        val accelData = NIDSensorHelper.getAccelerometerInfo()
         val currentHashValue = sequence?.toString()?.getSHA256withSalt()?.take(8)
 
         if (lastHashValue != currentHashValue) {
             lastHashValue = sequence?.toString()?.getSHA256withSalt()?.take(8)
-            NIDLog.d(msg = "Activity - after text ${sequence.toString()}")
+            logger.d(msg = "Activity - after text ${sequence.toString()}")
 
-            NeuroID.getInstance()?.dataStore
-                ?.saveEvent(
+            dataStore.saveEvent(
                     NIDEventModel(
                         type = INPUT,
-                        ts = ts,
                         tg = hashMapOf(
                             "attr" to getAttrJson(sequence.toString()),
                             "etn" to INPUT,
@@ -102,8 +91,6 @@ class NIDTextWatcher(
                         tgs = idName,
                         v = "S~C~~${sequence?.length}",
                         hv = sequence?.toString()?.getSHA256withSalt()?.take(8),
-                        gyro = gyroData,
-                        accel = accelData
                     )
                 )
         }
