+++ NeuroID/src/test/java/com/neuroid/tracker/NIDTextWatcherTests.kt
@@ -3,12 +3,11 @@ package com.neuroid.tracker
 import android.content.ClipData
 import android.content.ClipboardManager
 import android.text.Editable
+import com.neuroid.tracker.utils.NIDLogWrapper
 import com.neuroid.tracker.utils.NIDTextWatcher
 import io.mockk.every
 import io.mockk.mockk
-import io.mockk.unmockkAll
 import io.mockk.verify
-import org.junit.Before
 import org.junit.Test
 
 class NIDTextWatcherTests : NeuroIDClassUnitTests() {
@@ -20,10 +19,10 @@ class NIDTextWatcherTests : NeuroIDClassUnitTests() {
 
         val clipboardManager = mockk<ClipboardManager>()
         every { clipboardManager.primaryClip } returns clipData
-
         NeuroID.getInstance()?.setClipboardManagerInstance(clipboardManager)
-        setMockedDataStore()
-        val textWatcher = NIDTextWatcher("test", "myclass", "")
+        val dataStoreMock = setMockedDataStore()
+
+        val textWatcher = NIDTextWatcher(dataStoreMock, NIDLogWrapper(), "test", "myclass", "")
         textWatcher.onTextChanged("existing text", 0, 1, 0)
         verify(exactly = 0) { NeuroID.getInstance()?.dataStore?.saveEvent(any()) }
 
@@ -37,10 +36,9 @@ class NIDTextWatcherTests : NeuroIDClassUnitTests() {
         val clipboardManager = mockk<ClipboardManager>()
         every { clipboardManager.primaryClip } returns clipData
         NeuroID.getInstance()?.setClipboardManagerInstance(clipboardManager)
+        val dataStoreMock = setMockedDataStore()
 
-        setMockedDataStore()
-
-        val textWatcher = NIDTextWatcher(
+        val textWatcher = NIDTextWatcher(dataStoreMock, NIDLogWrapper(),
             "test", "myclass", "sgs".hashCode().toString()
         )
 
@@ -53,9 +51,9 @@ class NIDTextWatcherTests : NeuroIDClassUnitTests() {
     fun textWatcher_after_text_changed() {
         val clipboardManager = mockk<ClipboardManager>()
         NeuroID.getInstance()?.setClipboardManagerInstance(clipboardManager)
-        setMockedDataStore()
+        val dataStoreMock = setMockedDataStore()
 
-        val textWatcher = NIDTextWatcher("test", "myclass", "")
+        val textWatcher = NIDTextWatcher(dataStoreMock, NIDLogWrapper(),"test", "myclass", "")
         val seq = mockk<Editable>()
         every { seq.toString() } returns "changed text"
         every { seq.length } returns 12
@@ -73,9 +71,9 @@ class NIDTextWatcherTests : NeuroIDClassUnitTests() {
         every { clipboardManager.primaryClip } returns clipData
 
         NeuroID.getInstance()?.setClipboardManagerInstance(clipboardManager)
-        setMockedDataStore()
+        val dataStoreMock = setMockedDataStore()
 
-        val textWatcher = NIDTextWatcher(
+        val textWatcher = NIDTextWatcher(dataStoreMock, NIDLogWrapper(),
             "test", "myclass", "existing text".hashCode().toString()
         )
         textWatcher.onTextChanged("existing copied text", 0, 11, 11)
@@ -88,9 +86,9 @@ class NIDTextWatcherTests : NeuroIDClassUnitTests() {
     fun textWatcher_after_duplicate_text_changed() {
         val clipboardManager = mockk<ClipboardManager>()
         NeuroID.getInstance()?.setClipboardManagerInstance(clipboardManager)
-        setMockedDataStore()
+        val dataStoreMock = setMockedDataStore()
 
-        val textWatcher = NIDTextWatcher("test", "myclass", "")
+        val textWatcher = NIDTextWatcher(dataStoreMock, NIDLogWrapper(),"test", "myclass", "")
         val seq = mockk<Editable>()
         every { seq.toString() } returns "changed text"
         every { seq.length } returns 12
